Add some preflight checks to some potentially destructive operations

+++ cypress/support/commands.ts
@@ -378,16 +378,26 @@ Cypress.Commands.add(
     organisation: string = "CIEL",
     isCleanup: boolean = false
   ) => {
-    getAuthToken().then(authToken =>
+    getAuthToken().then(authToken => {
       cy.request({
-        method: "DELETE",
+        method: "GET",
         headers: {
           Authorization: authToken
         },
         url: `${apiUrl}/orgs/${organisation}/sources/${source}/`,
-        failOnStatusCode: !!!isCleanup
-      })
-    );
+      }).then(response => {
+        if (getUser() === response.body.created_by) {
+          cy.request({
+            method: "DELETE",
+            headers: {
+              Authorization: authToken
+            },
+            url: `${apiUrl}/orgs/${organisation}/sources/${source}/`,
+            failOnStatusCode: !!!isCleanup
+          });
+        }
+      });
+    });
   }
 );
 
@@ -454,16 +464,26 @@ Cypress.Commands.add(
 Cypress.Commands.add(
   "deleteOrganisation",
   (organisation: string, isCleanup: boolean = false) => {
-    getAuthToken().then(authToken =>
+    getAuthToken().then(authToken => {
       cy.request({
-        method: "DELETE",
+        method: "GET",
         headers: {
           Authorization: authToken
         },
         url: `${apiUrl}/orgs/${organisation}/`,
-        failOnStatusCode: !!!isCleanup
-      })
-    );
+      }).then(response => {
+        if (getUser() === response.body.created_by) {
+          cy.request({
+            method: "DELETE",
+            headers: {
+              Authorization: authToken
+            },
+            url: `${apiUrl}/orgs/${organisation}/`,
+            failOnStatusCode: !!!isCleanup
+          });
+        }
+      });
+    });
   }
 );
 
@@ -510,19 +530,32 @@ Cypress.Commands.add(
     concept_class: string = "Diagnosis"
   ) => {
     getAuthToken().then(authToken => {
-      cy.request({
-        method: "POST",
-        headers: {
-          Authorization: authToken
-        },
-        url: `${apiUrl}${source_url}concepts/`,
-        body: {
-          id: id,
-          concept_class: concept_class,
-          names: names,
-          datatype: "N/A"
-        }
-      });
+      cy
+        .request({
+          method: "GET",
+          headers: {
+            Authorization: authToken
+          },
+          url: `${apiUrl}${source_url}concepts/${id}`,
+          failOnStatusCode: false
+        })
+        .then(response => {
+          if (response.status !== 200) {
+            cy.request({
+              method: "POST",
+              headers: {
+                Authorization: authToken
+              },
+              url: `${apiUrl}${source_url}concepts/`,
+              body: {
+                id: id,
+                concept_class: concept_class,
+                names: names,
+                datatype: "N/A"
+              }
+            });
+          }
+        });
     });
 
     return cy.wrap(id);
